complete overhaul
a complete overhaul of the framework including sass and an expression
engine index file as well as the usual include structure if the site is
to be static. got rid of respond.js as from using it lately it doesn't
work that well with a "mobile first" approach as IE8 and below don't
handle that all too well. sorted the default js file to have fitvids
fittext and lettering my most common js and modernizr with common things
i use. general cleanup